update sklearn requirements and remove old code
