--- gplearn/skutils/README.rst
@@ -1,10 +1,0 @@
-skutils
-=======
-
-The modules contained within this folder are slightly modified versions of
-some key utility functions from scikit-learn 0.16.0 that gplearn depends upon.
-They reside here in order to maintain compatibility across different versions
-of scikit-learn.
-
-For original versions, see https://github.com/scikit-learn/scikit-learn/releases/tag/0.16.0
-
